[Platinum IV] Title: 날카로운 눈, Time: 328 ms, Memory: 24040 KB -BaekjoonHub

+++ 백준/Platinum/1637. 날카로운 눈/날카로운 눈.java
@@ -6,7 +6,7 @@ public class Main {
   public static void main(String[] args) throws IOException {
     BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
     int n = Integer.parseInt(br.readLine());
-    long[][] arrs = new long[n][3];
+    int[][] arrs = new int[n][3];
     for (int i = 0; i < n; i++) {
       StringTokenizer st = new StringTokenizer(br.readLine());
       arrs[i][0] = Integer.parseInt(st.nextToken());
@@ -20,7 +20,7 @@ public class Main {
       System.out.println(k + " " + count(k, arrs));
     }
   }
-  public static long two(long[][] arrs) {
+  public static long two(int[][] arrs) {
     long left = 0;
     long right = Long.MAX_VALUE;
 
@@ -28,10 +28,10 @@ public class Main {
       long mid = (left + right) / 2;
       long sum = 0;
       long count =0;
-      for(long[] arr : arrs){
-        long a = arr[0];
-        long c = arr[1];
-        long b = arr[2];
+      for(int[] arr : arrs){
+        int a = arr[0];
+        int c = arr[1];
+        int b = arr[2];
         if(a > mid) continue;
         if(c < a) continue;
         if(c < mid) count++;
@@ -46,12 +46,12 @@ public class Main {
     }
     return left;
   }
-  public static long count(long k, long[][] arrs){
+  public static long count(long k, int[][] arrs){
     long count =0 ;
-    for(long[] arr : arrs){
-      long a = arr[0];
-      long c = arr[1];
-      long b = arr[2];
+    for(int[] arr : arrs){
+      int a = arr[0];
+      int c = arr[1];
+      int b = arr[2];
       if(a > k || k > c) continue;
       if((k - a)%b == 0) count++;
     }
